requirements.txt
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
